Exemplar: Refactor to be under common.
